transferred files in kafkaStreamsEnergy to streamsKafka in order to re-create the POM without the maven template
