--- kafka/kafkaStreamsEnergy/src/main/java/kafkaStreams/JsonConnectSum.java
@@ -1,116 +1,0 @@
-package kafkaStreams;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.serialization.Serdes;
-import org.apache.kafka.streams.*;
-import org.apache.kafka.streams.kstream.Grouped;
-import org.apache.kafka.streams.kstream.KStream;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import java.io.IOException;
-import java.util.Properties;
-import java.util.concurrent.CountDownLatch;
-
-/**
- * A Kafka Streams application to calculate the cumulative sum of energy
- * of each geohash (location) that is providing data to the pipeline.
- */
-
-public class JsonConnectSum {
-
-    static String APPLICATION_ID = "application";
-    static String INPUT_TOPIC="fake_iot";
-    static String OUTPUT_TOPIC="streams-pipe-output";
-    static String broker = "127.0.0.1:9092";
-
-
-    public static void main(String[] args) {
-
-        // Kafka configuration
-        Properties props = new Properties();
-
-        props.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, broker);
-        props.put(StreamsConfig.APPLICATION_ID_CONFIG, APPLICATION_ID);
-        props.put(StreamsConfig.COMMIT_INTERVAL_MS_CONFIG, 1000);
-        props.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, Serdes.String().getClass());
-        props.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, Serdes.String().getClass());
-
-
-        // setting offset reset to earliest so that we can re-run the demo code with the same pre-loaded data
-        props.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
-
-
-        // create a logger for this class
-        Logger logger = LoggerFactory.getLogger(JsonConnectSum.class);
-
-        // initiate Kafka Streams Topology builder
-        final StreamsBuilder builder = new StreamsBuilder();
-
-        // Get the data from kafka topic as a key/value pair of strings
-        KStream<String, String> preJson = builder.stream(INPUT_TOPIC);
-
-        // map the topic key/value pair to a new key/value pair where key = geohash and value = energy
-        KStream<String, Double> geohashEnergy = preJson.map((key, value) -> {
-            KeyValue<String, Double> keyValue;
-
-
-            String jsonStr = value;
-            ObjectMapper mapper = new ObjectMapper();
-            try {
-                JsonNode jsonValues = mapper.readTree(jsonStr);
-
-                JsonNode geohash = jsonValues.get("geohash");
-                JsonNode energy = jsonValues.get("energyVal");
-                System.out.println(geohash+" "+energy);
-
-                String newKey = geohash.asText();
-                Double newValue = energy.asDouble();
-
-                keyValue = new KeyValue<>(newKey, newValue);
-                return keyValue;
-
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            keyValue = new KeyValue<String, Double>("", 0.0);
-
-            return keyValue;
-        });
-
-
-        // group by the geohash, and sum all energy values for that geohash, output to new kafka topic
-        geohashEnergy.groupByKey(Grouped.with(Serdes.String(), Serdes.Double()))
-                // Apply SUM aggregation
-                .reduce(Double::sum)
-                // Write to stream specified by outputTopic
-                .toStream().to(OUTPUT_TOPIC);
-
-
-        final Topology topology = builder.build();
-        final KafkaStreams streams = new KafkaStreams(topology, props);
-        final CountDownLatch latch = new CountDownLatch(1);
-
-
-        // attach shutdown handler to catch control-c
-        Runtime.getRuntime().addShutdownHook(new Thread("streams-shutdown-hook") {
-            @Override
-            public void run() {
-                streams.close();
-                latch.countDown();
-            }
-        });
-
-        try {
-            streams.cleanUp();
-            streams.start();
-            latch.await();
-        } catch (Throwable e) {
-            System.exit(1);
-        }
-        System.exit(0);
-    }
-
-}
