--- kafka/kafkaStreamsEnergy/src/resources/anomaly/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: kafkaStreams.MovingAverageAnomaly
-
