+++ kafka/streamsKafka/src/resources/moveavg/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: kafkaStreams.MovingAverageAnomaly
+
